feat: added functionality to look up missing image digests
Using Quay API, list repository tag objects which include
property 'manifest_digest' and set it up for all the operator
bundles that don't have it set up yet. This helps to match
recorded pull actions with identifiers in form of a manifest
digest only, which we couldn't match previously with some of
the operator bundles.