refactor: single --catalog-image argparse option with optional render skip
As it seems like a good idea to track with which catalogs and OCP versions
are operator bundles associated, the option --catalog-json-file is not
sufficient in providing the info. Therefore, this option was removed
and the --catalog-image option was extended by an optional argument.
Example usage: pullsar --catalog-image IMAGE [RENDERED_JSON_FILE];
If the rendered JSON file is provided, the script skips rendering
the IMAGE, uses RENDERED_JSON_FILE as source catalog of operators
and uses IMAGE just for database entry. If not provided, script
renders the IMAGE with 'opm render' as usual and continues from there.