feat: integrate PostgreSQL DB functionality for usage stats storage (#5)
1. Added support for connecting to PostgreSQL database and store results of a run by configuring '.env' file with credentials, see '.env.example'. Used tables schema (db/schema.py) aims to allow future queries for:
a) overall pull counts for each OCP version
b) overall pull counts for each catalog of operators
c) pull counts for each operator package
d) pull counts for each operator bundle
- all across chosen date range.

2. Refactored 'argparse' options:
a) As it seems like a good idea to track with which catalogs and OCP versions
are operator bundles associated, the option --catalog-json-file is not
sufficient in providing the info alone. Therefore, this option was removed
and the --catalog-image option was extended by an optional argument.
Example usage: pullsar --catalog-image IMAGE [RENDERED_JSON_FILE];
If the rendered JSON file is provided, the script skips rendering
the IMAGE, uses RENDERED_JSON_FILE as source catalog of operators
and uses IMAGE just for database entry (catalog and OCP version). If not provided,
script renders the IMAGE with 'opm render' as usual and continues from there.
b) added --dry-run/--test option, displaying the results on stdout but
but not storing them in the database.

3. Switch to working with DB friendly datetime.date objects instead of date strings.