feat: add export as CSV feature
1. Added backend API endpoints /export/csv.
2. Added backend API crud function get_all_items_for_export to retrieve items similarly
to the get_paginated_items but without the pagination.
3. Modified string date format in backend responses to be in YYYY-MM-DD, previously
it was 'Aug 19'-like, which should've been converted to on frontend, not on backend already.
Now formatted on frontend.
4. Added export button using new useExportCsv hook. The button is disabled
in case the date range is more than EXPORT_MAX_DAYS (currently 30 days). Also,
it's disabled immediately after exporting until the URL parameters change to prevent
repetetive API calls. User is informed about this by a tooltip on button hover.