feat: for worker, store repository pull logs for reuse optimization
When processing multiple OCP versions of the same catalog, we encounter the same
repository paths over and over. With the request to process many OCP versions,
optimization was really called for. With this simple change, pull logs are saved
globally for each processed repository path and reused upon repeated encounter
of that repository path instead of having to list the logs using same API requests again.