fix: worker jobs produce errors and take long (#12)
1. Fixed bug producing errors in logs:
For certified-operators and redhat-marketplace catalogs, when processing multiple OCP versions of the same catalog, the second and subsequent process was trying to send API requests to non-quay repositories - this was because when using the know_image_translations (translation of non-quay image to quay image is stored for repeated encounters), by mistake the old repo_path was used and this one was later used when making API requests. The errors in logs are now gone. Coincidentally, this probably made no difference to the results, as repeated encounter of a bundle means the bundle already has the pull count set, the wrong repo was skipped and the bundle's appearance in another catalog was still properly set to all the associated OCP versions.

2. Repository pull logs are now stored for future reuse:
When processing multiple OCP versions of the same catalog, we encounter the same
repository paths over and over. With the request to process many OCP versions,
optimization was really called for. With this simple change, pull logs are saved
globally for each processed repository path and reused upon repeated encounter
of that repository path instead of having to list the logs using same API requests again.

3. Repository tags and Pyxis images requested for a repository are also stored for reuse the same way.

4. Refactored: transform operator stats resolution into class with attribute cache:
- Renamed update_operator_usage_stats.py -> stats_resolver.py.
- Created new class OperatorUsageStatsResolver with attribute _cache of type CachedContext.
- Transformed all of the old module functions into methods of the new class.
- Initialized the class in main.
- Updated tests to reflect these changes.