refactor: resolution of PostgreSQL image to avoid permission issues
When using the image resolution within template, it seems that it needs to create some temporary objects
in the openshift namespace. To avoid the need for setting up writing permissions, the image is now resolved
just by reading the image from the API and passing it down to the template.

I have also simplified the template a bit and fixed database migration job behavior - indentation issues, time to live, removal of finished jobs.