fix: deployment to prod failing due to wrong group vars resolution (#16)
During deployment to production, despite Ansible playbook running with env=prod, the group vars were being resolved as if it was stage deployment. This resolved in quick failure due to trying to deploy to pullsar-stage namespace, which does not exist in the prod cluster.

I have fixed ansible/inventory/hosts.yml and run the workflow manually using workflow-dispatch from this branch to ensure the changes fix this and any other potential problem with deployment to prod. The deployment to prod has run successfully, therefore I consider this to be the final fix.